refactor(evaluations): extraire le composable useTeacherEvaluations (#28)
Tranche 2/3 de la décomposition de TeacherEvaluations.vue.

- useTeacherEvaluations.js (168 l.) : état + filtres + chargements
  (KLASSCI+LMS, cache classes/matières, fallback dashboard) + dérivés
  (evaluationsWithOnline / filteredEvaluations / stats /
  expiredWithoutOnlineCount via les fonctions pures de utils/evaluations).
- La vue consomme le composable par destructuration ; fetch/état/dérivés
  retirés (1800 -> 1654 l.). Mappers de statut gardés (usage template).
- Comportement préservé (mêmes appels, cache, dérivés).

Cumul décomposition : 1877 -> 1654 l. Reste tranche 3 (sous-composants à slots).

Tests 180/180, contract 28/28, build OK.
